fix test for Int32 machines

+++ test/test-permutations.jl
@@ -13,7 +13,7 @@ using SymPy
 using SymPy.Permutations
 using PyCall
 using Base.Test
-Base.range(i::Int64) = 0:(i-1)
+Base.range(i) = 0:(i-1)
 
 @testset "Permutations" begin
